tamer: asg::air: Remove unncessary vec![] usage in tests
This has been bothering me for quite a long time, and is just more test
cleanup before I introduce more.  I suspect this came from habit with the
previous Rust edition where `into_iter()` on arrays was a much more verbose
operation.

To be clear: this change isn't for performance.  It's about not doing
something silly when it's unnecessary, which also sets a bad example for
others.

There are many other tests in other modules that will need updating at some
point.

DEV-13163